requirements.txt
streamlit>=1.38.0
PyMuPDF==1.24.7
langchain-huggingface>=0.0.3
langchain-community>=0.0.38
faiss-cpu>=1.8.0
sentence-transformers>=2.3.1
pillow>=10.0.0
numpy>=1.24.0
